Refactor search box into its own component
